refactor(colyseus): split room to multiple files
